@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -33,6 +33,7 @@ dependencies:
   flutter_kakao_login: ^3.3.0
   provider: ^5.0.0
   shared_preferences: ^2.0.6
+  intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
