@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class NoticeDetailPage extends StatelessWidget {
   final String header, contents;
 
-  NoticeDetailPage({this.header, this.contents});
+  NoticeDetailPage({required this.header, required this.contents});
 
   @override
   Widget build(BuildContext context) {
